v7.5.0.2

+++ src/page-builder/Infrastructure/WordPress/WpAutopDisabler.php
@@ -14,7 +14,9 @@ final class WpAutopDisabler
 
     public function maybeDisable(): void
     {
-        if (is_singular() && $this->publicPageRenderPolicy->isReady(get_the_ID())) {
+        // The function get_the_ID() can return false when the wp hook has no current post.
+        // PublicPageRenderPolicy rejects the normalized 0 and keeps wpautop active.
+        if (is_singular() && $this->publicPageRenderPolicy->isReady((int) get_the_ID())) {
             remove_filter('the_content', 'wpautop');
         }
     }
